refactor: Move task to its own module
